Refactor structure and code cleanup
