+++ src/main/java/controller/Encryptor.java
@@ -1,4 +1,4 @@
-package com.davincecode.onlime;
+package controller;
 
 import java.math.BigInteger;
 import java.security.MessageDigest;
