+++ src/main/java/module-info.java
@@ -1,10 +1,12 @@
-module com.davincecode.onlime {
+module OnLimeChatter {
     requires javafx.controls;
     requires javafx.fxml;
     requires java.sql;
+    requires java.desktop;
     requires java.dotenv;
 
+    opens controller to javafx.fxml;
+    opens client to javafx.fxml;
 
-    opens com.davincecode.onlime to javafx.fxml;
-    exports com.davincecode.onlime;
+    exports client;
 }
