--- src/main/resources/com/davincecode/onlime/MainScreen.fxml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.Button?>
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.davincecode.onlime.MainScreen">
-   <children>
-      <VBox layoutX="50" layoutY="50" spacing="10">
-         <!-- ListView to display messages -->
-         <ListView fx:id="messageListView" prefHeight="250" prefWidth="500" />
-
-         <!-- HBox to contain the text field and send button -->
-         <HBox spacing="10">
-            <!-- TextField for typing messages -->
-            <TextField fx:id="messageTextField" onAction="#sendMessage" promptText="Type your message..." />
-
-            <!-- Button to send messages -->
-            <Button onAction="#sendMessage" text="Send" />
-         </HBox>
-      </VBox>
-   </children>
-</AnchorPane>
