+++ src/main/resources/views/MainScreen.fxml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<?import javafx.scene.layout.VBox?>
+<AnchorPane fx:id="ap_main" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="450.0" prefWidth="680.0" xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml">
+   <children>
+      <Label layoutX="307.0" layoutY="14.0" text="O F F L I M E" />
+      <ScrollPane fx:id="sp_user_list" layoutX="14.0" layoutY="38.0" prefHeight="329.0" prefWidth="216.0" />
+       <ScrollPane fx:id="sp_main" layoutX="244.0" layoutY="38.0" prefHeight="329.0" prefWidth="422.0" >
+           <content>
+               <VBox fx:id="vb_main_messages" prefHeight="325.0" prefWidth="415.0" />
+           </content>
+       </ScrollPane>
+      <TextField fx:id="tf_message" layoutX="244.0" layoutY="379.0" prefHeight="57.0" prefWidth="342.0" />
+      <Button fx:id="button_send" layoutX="597.0" layoutY="379.0" mnemonicParsing="false" text="Send">
+         <padding>
+            <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
+         </padding>
+      </Button>
+   </children>
+</AnchorPane>
